Fixed issue where class type prevented updates and assignments got accidentally marked as not done
PCR has renamed "projects" to "long term," which created errors since I
never prepared for spaces to be in an assignment's category name.
This update should fix that, as well as fixing an issue where
assignments would sometimes be accidently marked as not done.
If you see the word "project" at the beginning of a title in an
assignment in normal PCR, (which would invalidate my assumption that
this was a replacement) please tell me through the feedback form.