Updated PCR urls, other minor improvements
PCR has now changed the URL of almost every page.
I guess the "SC" in the new URL is somehow worse than the old "Student."

Anyways, Check PCR is not going to update itself since a 404 page isn't
going to have anything of interest on it.

This update should fix that as well as a few minor nuances such as the
retry button in the snackbar not working.