Add Google Analytics back
I've decided that being able to see the number of user who are using Check PCR is very valuable to me (it's interesting!). While I don't like that Google Analytics records so much data, it's the best thing I have at the moment, so I'm going to add it back. If anyone has it disabled already, this shouldn't change anything (that setting should still be respected).