Version 2.26.1: Yoink
If you're expecting this to be a bug fix, I'm so sorry. I can't do this anymore. It's like when you're such a good fisherman you catch all the salmon in the lake and all that's left are the sharks. I can't catch sharks, and that's why I never get any shark fin soup.

What I can do is add features. That is precisely what this update brings you: a feature. I highly recommend you check the settings menu after the update. So what can I say, except, you're welcome.